Refactor code for clarity and readability
The code in the CombinedCreatureCombinationTable.vue file has been refactored to improve readability and maintainability. Single line if-statements have been expanded into multiple lines to make the code easier to follow. A necessary condition switch has also been adjusted in the 'toggleShowAllColumns' function, which further enhances the structure and logic of the code.