Add new version of Tellurian to backend and update compressor script
This commit adds in a new version of Tellurian to the global test constants and mod schemas in the backend, labelled as version '2.10.0.3'. It also includes an update to compressor.ts script for better handling of file paths and improved logging outputs. Furthermore, it cleans up the 'inspectionProfiles' by removing unnecessary checks.