Enhanced modStore with fetchMods method and modified frontend elements
A fetchMods method was added to the modStore to fetch mods. The `modStore` was then imported and the `fetchMods()` action called in `app.vue`. Enums for CombinationAttributeNames and FilterOperators were created in both frontend and backend. The `getCombinationsRequestBody.ts` and `ModColumn.ts` types were modified, and the sorting type was added. Consistent usage of enum values was implemented in 'CombinationsTable.vue.