Updated package.json and improved environment variable handling
Updated package.json formatting and dependencies, which includes refining the build properties and adding more dependencies to the project. Also, the handling of environment variables has been improved in the serverEnvironment.ts file to default to production environment if not set. Further, 'globals.ts' and 'package-lock.json' were modified to align with the changes.